Move controller lease counter to its own package
