+++ pkg/component/controller/leasecounter/component.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 k0s authors
 // SPDX-License-Identifier: Apache-2.0
 
-package controller
+package leasecounter
 
 import (
 	"context"
